fix: add type to payloed file
